Fix style issues and install bunyan
