+++ package.json
@@ -36,6 +36,7 @@
   "version": "0.1.0",
   "devDependencies": {
     "@babel/preset-env": "^7.19.3",
+    "bunyan": "^1.8.15",
     "jest": "^29.1.2",
     "jest-environment-jsdom": "^29.1.2",
     "markdown-toc": "^1.2.0",
